Move types from backend to common package
Some other types were required in the new node plugin, so the best
way to approach this is moving those types to the common folder
instead, and use them both in the backend and node packages.
