+++ plugins/s3-viewer-common/src/index.ts
@@ -1,6 +1,8 @@
 export * from './cookie';
 export * from './types';
 export {
+  type BucketDetailsFilter,
+  type BucketDetailsFilters,
   type S3BucketDetailsPermission,
   S3_VIEWER_RESOURCE_TYPE,
   permissions,
