+++ plugins/s3-viewer-common/src/types.ts
@@ -34,3 +34,22 @@ export type FetchObjectResult = {
   downloadName: string;
   downloadUrl: string;
 };
+
+export type S3Platform = {
+  endpoint: string;
+  endpointName: string;
+  region: string;
+  credentials?: {
+    accessKeyId: string;
+    secretAccessKey: string;
+  };
+};
+
+export type BucketCredentials = S3Platform & {
+  bucket: string;
+};
+
+export type AllowedBuckets = {
+  platform: string;
+  buckets: string[];
+};
